pyproject.toml
[tool.poetry]
name = "bigbot"
version = "0.2.0"
description = "!bigbot"
authors = ["trav <heytrav205@gmail.com> jack <jack@flnpb.com>"]
readme = "README.md"
packages = [{include = "big_bot"}]

[tool.poetry.dependencies]
aiohttp = "3.7.4.post0"
async-timeout = "3.0.1"
attrs = "22.1.0"
certifi = "^2022.12.7"
chardet = "4.0.0"
idna = "3.4"
meson = "0.63.2"
multidict = "6.0.2"
python = "^3.10"
python-discord = "1.7.3"
python-dotenv = "0.21.0"
typing-extensions = "4.3.0"
yarl = "1.8.1"
discord-py = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
